move document page container
