+++ apps/scoutgame/components/common/DocumentPageContainer/components/Blockquote.tsx
