+++ apps/scoutgame/components/common/DocumentPageContainer/components/InfoCard.tsx
